#74 refactor: work dto 경로 수정
