+++ src/main/java/com/susanghan_guys/server/work/dto/request/DcaWorkSubmissionRequest.java
@@ -1,7 +1,8 @@
-package com.susanghan_guys.server.work.dto;
+package com.susanghan_guys.server.work.dto.request;
 
 import com.susanghan_guys.server.work.domain.type.Brand;
 import com.susanghan_guys.server.work.domain.type.Category;
+import com.susanghan_guys.server.work.dto.response.TeamMemberResponse;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
