+++ src/main/java/com/susanghan_guys/server/work/infrastructure/mapper/DcaWorkMapper.java
@@ -3,9 +3,7 @@ package com.susanghan_guys.server.work.infrastructure.mapper;
 import com.susanghan_guys.server.contest.domain.Contest;
 import com.susanghan_guys.server.user.domain.User;
 import com.susanghan_guys.server.work.domain.Work;
-import com.susanghan_guys.server.work.domain.type.Brand;
-import com.susanghan_guys.server.work.domain.type.Category;
-import com.susanghan_guys.server.work.dto.DcaWorkSubmissionRequest;
+import com.susanghan_guys.server.work.dto.request.DcaWorkSubmissionRequest;
 import org.springframework.stereotype.Component;
 
 @Component
