+++ src/main/java/com/susanghan_guys/server/work/infrastructure/mapper/YccWorkMapper.java
@@ -3,8 +3,7 @@ package com.susanghan_guys.server.work.infrastructure.mapper;
 import com.susanghan_guys.server.contest.domain.Contest;
 import com.susanghan_guys.server.user.domain.User;
 import com.susanghan_guys.server.work.domain.Work;
-import com.susanghan_guys.server.work.domain.type.FilesType;
-import com.susanghan_guys.server.work.dto.YccWorkSubmissionRequest;
+import com.susanghan_guys.server.work.dto.request.YccWorkSubmissionRequest;
 import org.springframework.stereotype.Component;
 
 @Component
