+++ src/main/java/com/susanghan_guys/server/work/presentation/WorkController.java
@@ -3,8 +3,8 @@ package com.susanghan_guys.server.work.presentation;
 import com.susanghan_guys.server.global.common.CommonResponse;
 import com.susanghan_guys.server.work.application.DcaWorkService;
 import com.susanghan_guys.server.work.application.YccWorkService;
-import com.susanghan_guys.server.work.dto.DcaWorkSubmissionRequest;
-import com.susanghan_guys.server.work.dto.YccWorkSubmissionRequest;
+import com.susanghan_guys.server.work.dto.request.DcaWorkSubmissionRequest;
+import com.susanghan_guys.server.work.dto.request.YccWorkSubmissionRequest;
 import com.susanghan_guys.server.work.presentation.swagger.WorkSwagger;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
