+++ src/main/java/com/susanghan_guys/server/work/presentation/swagger/WorkSwagger.java
@@ -1,20 +1,17 @@
 package com.susanghan_guys.server.work.presentation.swagger;
 
 import com.susanghan_guys.server.global.common.CommonResponse;
-import com.susanghan_guys.server.work.dto.DcaWorkSubmissionRequest;
-import com.susanghan_guys.server.work.dto.YccWorkSubmissionRequest;
-import com.susanghan_guys.server.work.dto.response.MyReportListResponse;
+import com.susanghan_guys.server.work.dto.request.DcaWorkSubmissionRequest;
+import com.susanghan_guys.server.work.dto.request.YccWorkSubmissionRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Encoding;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
